Typeset \alpha and \beta
Typeset the Beta distribution parameters \alpha and \beta to match as
they appear in the rest of the notebook.

In addition, also add a guard for N_0 in an attempt to help it be
properly typset. In the Towards Data Science post (viewed on 2019-05-16)
the 0 is not propley typset and is disjoint from the N.
https://towardsdatascience.com/https-medium-com-hankroark-finding-bayesian-legos-part1-b8aeb886afba